+++ pkg/api/response/lease.go
@@ -7,19 +7,8 @@ import (
 // CreateLeaseResponse creates an Lease Response based
 // on the provided Lease
 func CreateLeaseResponse(lease *db.Lease) *LeaseResponse {
-	return &LeaseResponse{
-		AccountID:                lease.AccountID,
-		PrincipalID:              lease.PrincipalID,
-		ID:                       lease.ID,
-		LeaseStatus:              lease.LeaseStatus,
-		CreatedOn:                lease.CreatedOn,
-		LastModifiedOn:           lease.LastModifiedOn,
-		BudgetAmount:             lease.BudgetAmount,
-		BudgetCurrency:           lease.BudgetCurrency,
-		BudgetNotificationEmails: lease.BudgetNotificationEmails,
-		LeaseStatusModifiedOn:    lease.LeaseStatusModifiedOn,
-		ExpiresOn:                lease.ExpiresOn,
-	}
+	response := LeaseResponse(*lease)
+	return &response
 }
 
 // LeaseResponse is the structured JSON Response for an Lease
@@ -35,16 +24,17 @@ func CreateLeaseResponse(lease *db.Lease) *LeaseResponse {
 // 	"BudgetNotificationEmails": ["usermsid@test.com", "managersmsid@test.com"]
 // }
 type LeaseResponse struct {
-	AccountID                string               `json:"accountId"`
-	PrincipalID              string               `json:"principalId"`
-	ID                       string               `json:"id"`
-	LeaseStatus              db.LeaseStatus       `json:"leaseStatus"`
-	LeaseStatusReason        db.LeaseStatusReason `json:"leaseStatusReason"`
-	CreatedOn                int64                `json:"createdOn"`
-	LastModifiedOn           int64                `json:"lastModifiedOn"`
-	BudgetAmount             float64              `json:"budgetAmount"`
-	BudgetCurrency           string               `json:"budgetCurrency"`
-	BudgetNotificationEmails []string             `json:"budgetNotificationEmails"`
-	LeaseStatusModifiedOn    int64                `json:"leaseStatusModifiedOn"`
-	ExpiresOn                int64                `json:"expiresOn"`
+	AccountID                string                 `json:"accountId"`
+	PrincipalID              string                 `json:"principalId"`
+	ID                       string                 `json:"id"`
+	LeaseStatus              db.LeaseStatus         `json:"leaseStatus"`
+	LeaseStatusReason        db.LeaseStatusReason   `json:"leaseStatusReason"`
+	CreatedOn                int64                  `json:"createdOn"`
+	LastModifiedOn           int64                  `json:"lastModifiedOn"`
+	BudgetAmount             float64                `json:"budgetAmount"`
+	BudgetCurrency           string                 `json:"budgetCurrency"`
+	BudgetNotificationEmails []string               `json:"budgetNotificationEmails"`
+	LeaseStatusModifiedOn    int64                  `json:"leaseStatusModifiedOn"`
+	ExpiresOn                int64                  `json:"expiresOn"`
+	Metadata                 map[string]interface{} `json:"metadata"`
 }