+++ pkg/db/model.go
@@ -20,18 +20,19 @@ type Account struct {
 // Lease is a type corresponding to a Lease
 // table record
 type Lease struct {
-	AccountID                string            `json:"AccountId"`                // AWS Account ID
-	PrincipalID              string            `json:"PrincipalId"`              // Azure User Principal ID
-	ID                       string            `json:"Id"`                       // Lease ID
-	LeaseStatus              LeaseStatus       `json:"LeaseStatus"`              // Status of the Lease
-	LeaseStatusReason        LeaseStatusReason `json:"LeaseStatusReason"`        // Reason for the status of the lease
-	CreatedOn                int64             `json:"CreatedOn"`                // Created Epoch Timestamp
-	LastModifiedOn           int64             `json:"LastModifiedOn"`           // Last Modified Epoch Timestamp
-	BudgetAmount             float64           `json:"BudgetAmount"`             // Budget Amount allocated for this lease
-	BudgetCurrency           string            `json:"BudgetCurrency"`           // Budget currency
-	BudgetNotificationEmails []string          `json:"BudgetNotificationEmails"` // Budget notification emails
-	LeaseStatusModifiedOn    int64             `json:"LeaseStatusModifiedOn"`    // Last Modified Epoch Timestamp
-	ExpiresOn                int64             `json:"ExpiresOn"`                // Lease expiration time as Epoch
+	AccountID                string                 `json:"AccountId"`                // AWS Account ID
+	PrincipalID              string                 `json:"PrincipalId"`              // Azure User Principal ID
+	ID                       string                 `json:"Id"`                       // Lease ID
+	LeaseStatus              LeaseStatus            `json:"LeaseStatus"`              // Status of the Lease
+	LeaseStatusReason        LeaseStatusReason      `json:"LeaseStatusReason"`        // Reason for the status of the lease
+	CreatedOn                int64                  `json:"CreatedOn"`                // Created Epoch Timestamp
+	LastModifiedOn           int64                  `json:"LastModifiedOn"`           // Last Modified Epoch Timestamp
+	BudgetAmount             float64                `json:"BudgetAmount"`             // Budget Amount allocated for this lease
+	BudgetCurrency           string                 `json:"BudgetCurrency"`           // Budget currency
+	BudgetNotificationEmails []string               `json:"BudgetNotificationEmails"` // Budget notification emails
+	LeaseStatusModifiedOn    int64                  `json:"LeaseStatusModifiedOn"`    // Last Modified Epoch Timestamp
+	ExpiresOn                int64                  `json:"ExpiresOn"`                // Lease expiration time as Epoch
+	Metadata                 map[string]interface{} `json:"Metadata"`                 // Arbitrary key-value metadata to store with lease object
 }
 
 // Timestamp is a timestamp type for epoch format